feat: move Firebase initialization from repository folder to infra shared folder
